moved new candidateSelection code to its own files (and sub namespace)
