+++ src/conv/heuristics/ai_heuristics.cpp
@@ -24,11 +24,18 @@
  *
  *******************************************************************************/
 
+// This file implements AI-based heuristics for convolution solver and kernel tuning.
+// Sections:
+//   1. Includes
+//   2. Common utilities
+//   3. Immediate mode AI models
+//   4. Kernel tuning AI models
+
+#include <numeric> // For std::inner_product
 #include <miopen/conv/heuristics/ai_heuristics.hpp>
 #if MIOPEN_ENABLE_AI_IMMED_MODE_FALLBACK || MIOPEN_ENABLE_AI_KERNEL_TUNING
 #include <fdeep/fdeep.hpp>
 #include <miopen/filesystem.hpp>
-#include <numeric> // For std::inner_product
 
 namespace miopen {
 namespace ai {
@@ -787,399 +794,6 @@ bool ModelSetParams(const std::string& arch,
     MIOPEN_LOG_I2("KTN ran for " << duration.count() << " micro-seconds");
     return true;
 }
-
-// code for new-style AI heuristics for kernel tuning
-
-// Implement the constructor as a separate method
-CandidateSelectionMetadata::CandidateSelectionMetadata(const std::string& arch, const std::string& solver)
-{
-    const nlohmann::json metadata =
-        common::LoadJSON(GetSystemDbPath() / (arch + "_" + solver + "_metadata.tn.model"));
-
-    // Input/output parameter order
-    input_params = metadata.value("input_params", std::vector<std::string>{});
-    output_params = metadata.value("output_params", std::vector<std::string>{});
-
-    // Build index maps for fast lookup
-    for(size_t i = 0; i < input_params.size(); ++i)
-        input_param_indices[input_params[i]] = i;
-    for(size_t i = 0; i < output_params.size(); ++i)
-        output_param_indices[output_params[i]] = i;
-
-    // Encodings
-    if(metadata.contains("encodings"))
-    {
-        feature_encodings = metadata["encodings"].value("inputs", decltype(feature_encodings){});
-        sequence_encodings = metadata["encodings"].value("outputs", decltype(sequence_encodings){});
-    }
-
-    // Decodings
-    if(metadata.contains("decodings") && metadata["decodings"].contains("outputs"))
-    {
-        sequence_decodings = metadata["decodings"]["outputs"]
-            .get<std::unordered_map<std::string, std::unordered_map<std::string, std::string>>>();
-    }
-
-    // Constants
-    if(metadata.contains("constants"))
-    {
-        constants_features = metadata["constants"].value("inputs", decltype(constants_features){});
-        constants_sequence = metadata["constants"].value("outputs", decltype(constants_sequence){});
-    }
-}
-
-// Implement the member functions as separate methods
-size_t CandidateSelectionMetadata::GetInputParamIndex(const std::string& name) const
-{
-    auto it = input_param_indices.find(name);
-    if(it == input_param_indices.end())
-        MIOPEN_THROW("Input parameter not found: " + name);
-    return it->second;
-}
-
-size_t CandidateSelectionMetadata::GetOutputParamIndex(const std::string& name) const
-{
-    auto it = output_param_indices.find(name);
-    if(it == output_param_indices.end())
-        MIOPEN_THROW("Output parameter not found: " + name);
-    return it->second;
-}
-
-std::optional<std::string> CandidateSelectionMetadata::GetInputConstant(const std::string& name) const
-{
-    auto it = constants_features.find(name);
-    if(it != constants_features.end())
-        return it->second;
-    return std::nullopt;
-}
-
-std::optional<std::string> CandidateSelectionMetadata::GetOutputConstant(const std::string& name) const
-{
-    auto it = constants_sequence.find(name);
-    if(it != constants_sequence.end())
-        return it->second;
-    return std::nullopt;
-}
-
-std::vector<size_t> CandidateSelectionMetadata::GetConstantInputIndices() const
-{
-    std::vector<size_t> indices;
-    for(const auto& [name, value] : constants_features)
-    {
-        auto it = input_param_indices.find(name);
-        if(it != input_param_indices.end())
-            indices.push_back(it->second);
-    }
-    std::sort(indices.begin(), indices.end());
-    return indices;
-}
-
-std::vector<size_t> CandidateSelectionMetadata::GetConstantOutputIndices() const
-{
-    std::vector<size_t> indices;
-    for(const auto& [name, value] : constants_sequence)
-    {
-        auto it = output_param_indices.find(name);
-        if(it != output_param_indices.end())
-            indices.push_back(it->second);
-    }
-    std::sort(indices.begin(), indices.end());
-    return indices;
-}
-
-class CandidateSelectionModel
-{
-public:
-    CandidateSelectionMetadata metadata;
-
-    CandidateSelectionModel(const std::string& arch, const std::string& solver)
-        : metadata(arch, solver),
-          input_encoder(fdeep::load_model(
-              InputEncoderPath(arch, solver), true, fdeep::dev_null_logger)),
-          kernel_config_encoder(fdeep::load_model(
-              KernelConfigEncoderPath(arch, solver), true, fdeep::dev_null_logger))
-    {
-    }
-    virtual ~CandidateSelectionModel() = default;
-
-    /**
-     * Encode input features into a context vector
-     * @param features Input problem features
-     */
-    fdeep::tensors EncodeInputFeatures(const std::vector<float>& features) const
-    {
-        // Remove constant parameters from features
-        std::vector<size_t> drop_indices = metadata.GetConstantInputIndices();
-        std::vector<float> filtered_features;
-        filtered_features.reserve(features.size() - drop_indices.size());
-
-        for(size_t i = 0, j = 0; i < features.size(); ++i)
-        {
-            if(j < drop_indices.size() && i == drop_indices[j])
-            {
-                ++j; // skip this index
-                continue;
-            }
-            filtered_features.push_back(features[i]);
-        }
-
-        fdeep::tensor input_tensor = fdeep::tensor(fdeep::tensor_shape(filtered_features.size()), filtered_features);
-        return input_encoder.predict({input_tensor});
-    }
-
-    /**
-     * Encode kernel configuration candidates
-     * @param encoded_candidates Encoded kernel parameter candidates
-     */
-    fdeep::tensors
-    EncodeKernelConfigs(const std::vector<std::vector<float>>& encoded_candidates) const
-    {
-        // Drop columns corresponding to constant output parameters
-        std::vector<size_t> drop_indices = metadata.GetConstantOutputIndices();
-
-        std::vector<std::vector<float>> filtered_candidates;
-        filtered_candidates.reserve(encoded_candidates.size());
-
-        for(const auto& candidate : encoded_candidates)
-        {
-            std::vector<float> filtered;
-            filtered.reserve(candidate.size() - drop_indices.size());
-            for(size_t i = 0, j = 0; i < candidate.size(); ++i)
-            {
-                if(j < drop_indices.size() && i == drop_indices[j])
-                {
-                    ++j; // skip this index
-                    continue;
-                }
-                filtered.push_back(candidate[i]);
-            }
-            filtered_candidates.push_back(filtered);
-        }
-
-        // Flatten candidates matrix for tensor input
-        std::vector<float> flattened_candidates;
-        for(const auto& candidate : filtered_candidates)
-        {
-            flattened_candidates.insert(
-                flattened_candidates.end(), candidate.begin(), candidate.end());
-        }
-
-        if(filtered_candidates.empty() || filtered_candidates[0].empty())
-        {
-            MIOPEN_THROW(miopenStatusInternalError,
-                         "Empty candidates provided to kernel config encoder");
-        }
-
-        fdeep::tensor candidates_tensor = fdeep::tensor(
-            fdeep::tensor_shape(encoded_candidates.size(), encoded_candidates[0].size()),
-            flattened_candidates);
-
-        return kernel_config_encoder.predict({candidates_tensor});
-    }
-
-    /**
-     * Select best candidate from encoded features and kernel configs
-     * @param encoded_features Encoded input features (1D tensor)
-     * @param encoded_configs Encoded kernel configurations (2D tensor)
-     */
-    int SelectBestCandidate(const fdeep::tensors& encoded_features,
-                            const fdeep::tensors& encoded_configs) const
-    {
-        const auto& feature_vec = encoded_features[0].to_vector();
-        const auto& config_mat = encoded_configs[0].to_vector();
-        const auto config_tensor  = encoded_configs[0];
-        
-        // Get dimensions from the 2D tensor shape
-        // For a 2D tensor, we need the first dimension (number of candidates)
-        // and second dimension (feature dimension)
-        const auto num_candidates = config_tensor.shape().volume() / feature_vec.size();
-        const auto feature_dim = feature_vec.size();
-        
-        // Verify dimensions are consistent
-        if(config_mat.size() != num_candidates * feature_dim)
-        {
-            MIOPEN_THROW(miopenStatusInternalError,
-                         "Inconsistent tensor dimensions in SelectBestCandidate");
-        }
-
-        std::vector<float> selection_scores(num_candidates, 0.0f);
-
-        // Calculate dot product between each candidate config and the feature vector
-        for(std::size_t i = 0; i < num_candidates; ++i)
-        {
-            selection_scores[i] = std::inner_product(
-                config_mat.begin() + i * feature_dim,
-                config_mat.begin() + (i + 1) * feature_dim,
-                feature_vec.begin(),
-                0.0f);
-        }
-
-        // Return index of candidate with highest score
-        return static_cast<int>(std::max_element(selection_scores.begin(), selection_scores.end()) -
-                               selection_scores.begin());
-    }
-
-private:
-    const fdeep::model input_encoder;
-    const fdeep::model kernel_config_encoder;
-
-    static std::string InputEncoderPath(const std::string& arch, const std::string& solver)
-    {
-        const auto path = GetSystemDbPath() / (arch + "_" + solver + "_input_encoder.tn.model");
-        if(!fs::exists(path))
-            MIOPEN_THROW(miopenStatusInternalError, "Unable to load input encoder file: " + path);
-        return path.string();
-    }
-
-    static std::string KernelConfigEncoderPath(const std::string& arch, const std::string& solver)
-    {
-        const auto path =
-            GetSystemDbPath() / (arch + "_" + solver + "_kernel_config_encoder.tn.model");
-        if(!fs::exists(path))
-            MIOPEN_THROW(miopenStatusInternalError,
-                         "Unable to load kernel config encoder file: " + path);
-        return path.string();
-    }
-};
-
-/**
- * Get the candidate selection model for given architecture and solver
- */
-std::shared_ptr<CandidateSelectionModel> GetCandidateSelectionModel(const std::string& arch,
-                                                                    const std::string& solver)
-{
-    static std::map<std::string, std::shared_ptr<CandidateSelectionModel>> models;
-    std::string key = arch + "_" + solver;
-    auto it         = models.find(key);
-    if(it == models.end())
-    {
-        std::shared_ptr<CandidateSelectionModel> model =
-            std::make_shared<CandidateSelectionModel>(arch, solver);
-        models[key] = model;
-        return model;
-    }
-    else
-    {
-        return it->second;
-    }
-}
-
-// Forward declaration
-std::vector<std::vector<float>>
-EncodeKernelParams(const std::vector<std::vector<std::string>>& valid_kernel_params,
-                   const CandidateSelectionMetadata& metadata);
-
-/**
- * Select the best candidate kernel parameters using the new candidate selection approach
- *
- * This function takes in a set of valid kernel parameters and features, encodes them,
- * and runs inference on the KernelTuningNet model to select the best candidate.
- *
- * @param arch GPU Architecture
- * @param solver Solver
- * @param direction Convolution Direction
- * @param features Input features for KernelTuningNet model
- * @param valid_kernel_params Valid kernel parameters to choose from
- */
-
-int ModelSelectBestCandidate(const std::string& arch,
-                             const std::string& solver,
-                             const std::vector<float>& features,
-                             const std::vector<std::vector<std::string>>& valid_kernel_params)
-{
-    try
-    {
-        auto model = GetCandidateSelectionModel(arch, solver);
-
-        // Encode string parameters to floats
-        auto encoded_candidates =
-            EncodeKernelParams(valid_kernel_params, model->metadata);
-
-        if(encoded_candidates.empty())
-        {
-            MIOPEN_LOG_W("No valid encoded candidates available");
-            return -1;
-        }
-
-        // Encode input features
-        auto encoded_features = model->EncodeInputFeatures(features);
-
-        // Encode kernel configurations
-        auto encoded_configs = model->EncodeKernelConfigs(encoded_candidates);
-
-        // Select best candidate
-        int best_idx = model->SelectBestCandidate(encoded_features, encoded_configs);
-
-        if(best_idx >= 0 && best_idx < static_cast<int>(valid_kernel_params.size()))
-        {
-            return best_idx;
-        }
-        else
-        {
-            MIOPEN_LOG_W("Invalid candidate index returned: " << best_idx);
-            return -1;
-        }
-    }
-    catch(const miopen::Exception& ex)
-    {
-        MIOPEN_LOG_I2("[Warning] Candidate selection model failed: " << ex.what());
-        return -1;
-    }
-    catch(const std::exception& ex)
-    {
-        MIOPEN_LOG_I2(
-            "[Warning] Candidate selection model failed with std exception: " << ex.what());
-        return -1;
-    }
-}
-
-// Helper function to encode kernel parameters
-std::vector<std::vector<float>>
-EncodeKernelParams(const std::vector<std::vector<std::string>>& valid_kernel_params,
-                   const CandidateSelectionMetadata& metadata)
-{
-    std::vector<std::vector<float>> encoded_candidates;
-    const auto& output_params = metadata.output_params;
-    const auto& sequence_encodings = metadata.sequence_encodings;
-
-    for(const auto& candidate : valid_kernel_params)
-    {
-        std::vector<float> encoded;
-        for(size_t i = 0; i < candidate.size(); ++i)
-        {
-            const std::string& param_name = output_params[i];
-            const std::string& param_value = candidate[i];
-
-            // Try to encode using the encoding map if available
-            auto enc_it = sequence_encodings.find(param_name);
-            if(enc_it != sequence_encodings.end())
-            {
-                const auto& value_map = enc_it->second;
-                auto val_it = value_map.find(param_value);
-                if(val_it != value_map.end())
-                {
-                    encoded.push_back(static_cast<float>(val_it->second));
-                    continue;
-                }
-            }
-
-            // If not found in encoding map, try to convert to float directly
-            try
-            {
-                encoded.push_back(std::stof(param_value));
-            }
-            catch(const std::exception&)
-            {
-                // Unknown parameter value, use sentinel
-                encoded.push_back(-1.0f);
-            }
-        }
-        encoded_candidates.push_back(encoded);
-    }
-
-    return encoded_candidates;
-}
-
 } // namespace tuning
 #endif // MIOPEN_ENABLE_AI_KERNEL_TUNING
 } // namespace ai
